(PC-11993) feat(identityCheck): update snapshot + extract translation
